# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SecureAssess Platform API",
    "version": "2.1.0",
    "description": "Authorized non-destructive application security assessment API"
  },
  "servers": [{ "url": "http://localhost:3847" }],
  "paths": {
...
    "/api/scan": {
      "post": {
        "summary": "Start assessment",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["targetUrl", "authorized"],
                "properties": {
                  "targetUrl": { "type": "string", "format": "uri" },
                  "authorized": { "type": "boolean" },
                  "projectName": { "type": "string" },
                  "projectId": { "type": "string" },
                  "username": { "type": "string" },
                  "password": { "type": "string" },
                  "securityTypes": { "type": "array", "items": { "type": "string" } },
                  "mode": {
                    "type": "string",
                    "enum": ["passive", "active-safe", "authenticated"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Job started" },
          "400": { "description": "Validation / target policy" }
        }
      }
    },
    "/api/scan/{jobId}": {
      "get": {
        "summary": "Get scan job status",
        "parameters": [
          { "name": "jobId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Job" }, "404": { "description": "Not found" } }
      },
      "delete": {
        "summary": "Cancel scan job",
        "parameters": [
          { "name": "jobId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Cancel requested" } }
      }
    },
...
    "/api/scans/compare": {
      "get": {
        "summary": "Compare two scans or latest project baseline",
        "parameters": [
          { "name": "baselineId", "in": "query", "schema": { "type": "string" } },
          { "name": "currentId", "in": "query", "schema": { "type": "string" } },
          { "name": "projectId", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Comparison" } }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  }
}